Clarify usage of Doc and doc to be more consistent
I was confused by this tutorial's inconsistent use of `Doc` and `doc`. I figured out that `Doc` refers to the type of object, and `doc` is an (arbitrary) variable name assigned to an instance of a `Doc` object. But as written, the tutorial refers to "**the** `doc`" when it really means "the **`Doc`**". Use of an article only makes sense when referring to the object, `Doc`, not when referring to the instance, `doc`. 

I have made a number of proposed changes to clarify the references. References to the instance, `doc`, should have no articles, because they refer to an individual thing, not a class of things (we wouldn't say "the Jane" or "a Jane" except in very specific situations). References to the object, `Doc`, should use an article and should be capitalized. I have also added a few phrases here and there to clarify how `doc`, a `Doc`, is created by use of `nlp`.